Split off module `Primitive.Types.Tx.TxOut.Gen`.
