+++ lib/wallet/test/unit/Cardano/Wallet/Primitive/ModelSpec.hs
@@ -100,9 +100,11 @@ import Cardano.Wallet.Primitive.Types.Tx
     , txScriptInvalid
     )
 import Cardano.Wallet.Primitive.Types.Tx.Gen
-    ( genTx, genTxOut, genTxScriptValidity, shrinkTx, shrinkTxOut )
+    ( genTx, genTxScriptValidity, shrinkTx )
 import Cardano.Wallet.Primitive.Types.Tx.TxIn.Gen
     ( genTxIn, shrinkTxIn )
+import Cardano.Wallet.Primitive.Types.Tx.TxOut.Gen
+    ( genTxOut, shrinkTxOut )
 import Cardano.Wallet.Primitive.Types.Tx.TxSeq
     ( TxSeq )
 import Cardano.Wallet.Primitive.Types.Tx.TxSeq.Gen
