+++ lib/wallet/test/unit/Cardano/Wallet/Shelley/Compatibility/LedgerSpec.hs
@@ -25,10 +25,10 @@ import Cardano.Wallet.Primitive.Types.TokenQuantity.Gen
     ( genTokenQuantityFullRange, shrinkTokenQuantityFullRange )
 import Cardano.Wallet.Primitive.Types.Tx
     ( TxIn (..) )
-import Cardano.Wallet.Primitive.Types.Tx.Gen
-    ( genTxOutCoin, shrinkTxOutCoin )
 import Cardano.Wallet.Primitive.Types.Tx.TxIn.Gen
     ( genTxIn, shrinkTxIn )
+import Cardano.Wallet.Primitive.Types.Tx.TxOut.Gen
+    ( genTxOutCoin, shrinkTxOutCoin )
 import Cardano.Wallet.Shelley.Compatibility.Ledger
     ( Convert (..) )
 import Data.Proxy
